Populate CONTRIBUTING.md
Guidelines for contributions that the team and any volunteers are now
in place.  They are borrowed and adapted directly from the Open Tech
Strategies PSM project (http://github.com/solutionguidance/psm).

The CONTRIBUTING.md also references a few other resources that are
not yet populated (such as TESTING.md, etc).  Placeholder files have
been added so as not to cause conflict.

Issue #14 Prepare the repository for contributions