Add Code of Conduct
This CoC was swiped directly from the [Open Tech Strategies COC](https://github.com/SolutionGuidance/psm/blob/master/docs/CODE_OF_CONDUCT.md)
but it seems to be pretty comprehensive.  Right now the Beastmaster is
listed as the person to reach out to in the event of a CoC violation,
though we may want to create an alias at BIFFUD that goes to the entire
Elected Overlord team for things like this.

Issue #14 - Prepare the repository for contributions