Add React skeleton (#24)
Add React skeleton and installation instructions

This commit contains the unmodified result of running create-react-app
in order to give us a baseline skeleton to build from going forward.  
There isn't much to the application yet, but in the spirit of documenting
as we go this includes documentation for kicking off what is there.