Set up sequelize
We added sequelize as a dependency already, but didn't put it to use.
This commit contains the byproducts of sequelize init, with a slight
modification to make sure that our env.js is loaded correctly.

Note: I'm using this [guide](https://scotch.io/tutorials/getting-started-with-node-express-and-postgres-using-sequelize)
for setting things up, since it seems reasonably close to what we're
trying to accomplish here.

Issue #26 Set up a database system