Add and configure ESLint
ESLint is a tool that magically tells us how bad we are at programming.
It's great! We have configured it to use the airbnb defaults, which is
just to say that we're going to be yelled at constantly.

The goal is to make it as difficult as possible to contribute to this
project. We want our contributors to feel bad about themselves, and
airbnb's linter has been optimized for exactly this.

For more information check out:

- https://eslint.org/docs/user-guide/getting-started
- https://github.com/airbnb/javascript/tree/master/packages/eslint-config-airbnb

Issue #32 Set up linting

Co-authored-by: Carolyn Rupar <crupar@gmail.com>