Add express
Express is a minimalist web framework for Node.

We are going to use express to serve our API endpoints.  It should be
noted that express will not be used for serving user facing content.

This commit doesn't do much to set up things like routes; it simply adds
the dependency and the baseline configuration setup.

Config introduced in this commit:

- API_PORT :: this is the port number that express will use to serve the
    API.

Issue #29 Add Express

Co-authored-by: Carolyn Rupar <crupar@gmail.com>