Add graphql dependencies
I'm pretty sure GraphQL just gives everybody complete read / write
access to your database, or something.  I don't know, everybody seems to
think it's great so it must be!  You can read more about it
[here](https://graphql.org/).

Meanwhile, this commit installs some basic backages for graphql.  The
first is graphql itself.  The second is a popular tool that makes it
easy to surface the grpahQL api over HTTP.

This also makes a modification to the babel pacakges we are using, since
some of the graphql scaffolding is going to leverage things like
destructuring which apparently requires the stage-2 preset.

Finally, it includes the pg package which should have been included in
the earlier sequelize setup but was missed.  This allows Sequelize to
actually interact with Postgresql.

Issue #36 Set up GraphQL / API