GraphQL skeleton
This commit introduces GraphQL to our stack, and also creates an example
schema field (though it does not have a corresponding model and is not
actually functional).

It also exposes the API on '/graphql' and a nifty interface at
'/graphiql' that makes it possible to test that API with a UI.

The scaffolding is inspired by [this
repository](https://github.com/atulmy/fullstack-graphql).  Once we have
a real schema entry we should delete the example one.

Issue #36 Set up GraphQL / API