Add five sample credible_content seed items
These were manually "scraped" but will hopefully reflect the same kind
of format that will come from the scraper once it is finished.
Specifically, it is going to mirror the content on the fact checking
sites, including paragraph tags, links, spans, etc.

If we ultimately decide to try to strip those things we will come back
and edit this seed list