Fixed credible content model
The description type was incorrectly a String, which isn't long enough.
It should have been TEXT which can hold a lot more data.

It is normally NOT OK to edit migrations, but since we are just getting
started and nobody is using this but us I'm going to take the flack
rather than create a tiny patch migration this early in the game.

You are going to have to manually run:

```
$> sequelize db:migrate:undo
$> sequelize db:migrate
```

which will delete your existing database and run the migrations again