Set up Parcel
We're using Parcel as a bundler for React.  This commit brings babel
where we need it to be, and sets up the package hooks for running
parcel.  For now it is really just configured for development -- the
compiled react will be served on port 1234.

Eventually we will want to have this stuff compile and get served from
the server directly...  At least I think that's what we'll want!

Issue #41