Populate navigation
This wires up the core navigation on the nav bar.  Everything is ugly
and nothing actually works, but the props and state changes propogate
across the entire application.  Review changes, new document, and the
about pane toggle (and closing the about pane) now trigger.

The new document button will eventually render a modal to confirm that
the user actually wants to wipe away their current state.  This is not
top priority so I won't build it out now.

Issue #45