Add editor textarea
The unstyled and baseline functional for the language checker editor
area.  This doesn't handle highlighting yet, but it does propogate state
updates through the app when text is edited.

Eventually we may find that we need to trigger state change less
frequently or otherwise be smarter about when to look up new flagged
phrases so it doesn't happen on every single keystroke, but this is
a fine place to start and may end up being optimally responsive.

Issue #52