Add dummy phrase detection
The highlights will be done using an approach similar to the one
outlined [over
here](https://codersblock.com/blog/highlight-text-inside-a-textarea/).

Essentially, we will have a div that mirrors the content of the editor
textare, except in the mirrored version the text is hidden and the text
gets highlights applied.

This commit is the first step towards that dream; specifically, it scans
the text that has been entered for any flagged phrases and then (for
now) wraps it in a bold tag.  Since react is react, we can't just use
string replace to accomplish it.  That's why this introduces a new
dependencey called `react-string-replace` which is "needed" in order to
insert jsx into a string.

Next step will be to have it insert an actual component instead of
a strong tag.

Issue #52