Add FlaggedPhrase component
When a phrase has been flagged we don't just want to make it bold, we
want it to become interactive.  The FlaggedPhrase component captures
that interactivity and also invokes styled components for the first time
in this project in order to have the background adjust color based on
the partisan counts.

This also passes selectd stat, though that information isn't yet used to
animate selection and deselection.

Issue #52