Add selected phrase detection
When the editor cursor is sitting on a selected phrase we want that
phrase to be identified as selected in terms of react state.

It is possible this will be a little slow, especially when we start
working with larger amounts of text or flags.  Optimization later
I suppose.

Originally I was going to try to find a way to have it detect "click"
events on the highlight layer; however, since the user can select text
using the keyboard it made more sense to use this more flexible approach.

Issue #52