Add flag selection animation
When a flag is selected, we want that to be visually represented with
a change in state across ALL instances of that selected flag.  In order
to do this without screwing everything up we needed to separate the
background color from the flag content.

This commit also removes some accidental cruft that snuck into the
FlaggedPhrase component.

Issue #52